NYD_Hackathon_2025/iks-rag-pipelines/requirements.txt
python-dotenv
groq
numpy<2
pandas
qdrant-client
spacy
better-profanity 
together